app/backend/pyproject.toml
[project]
name = "waste-classifier-backend"
version = "0.1.0"
description = "FastAPI backend for WasteClassifier MLOps web app"
authors = [{name = "Grzegorz Gomza", email = "your.email@example.com"}]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "torch",
    "torchvision",
    "xgboost",
    "numpy",
    "pillow",
    "scikit-image",
    "scipy",
    "pydantic",
    "python-multipart",
    "pydantic-settings",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]
